restructuring gui scripts
